+++ django_ctb/services/sync.py
@@ -68,7 +68,7 @@ class ProjectVersionBomService:
         _part = self._get_matching_parts(row=row).first()
         return _part
 
-    def _sync_footprints(serf, footprint_refs, *, project_part):
+    def _sync_footprints(self, footprint_refs, *, project_part):
         # get rid of any outdated/altered refs
         models.ProjectPartFootprintRef.objects.filter(
             project_part=project_part